// src/app/models/admin/dangkyhoatdongkhoangsan/dangkytanthu/dktanthuloaikhoangsan.model.ts
export class OutputDkTanThuLoaiKhoangSan {
  idtanthuloaikhoangsan: string;
  tenkhoangsan: string;
  idloaikhoangsan: string;
  tenloaikhoangsan: string;
  idnhomkhoangsan: string;
  tennhomkhoangsan: string;
  iddangkytanthu: string;
  truluong: number;
  donvitruluong: string;
  siteid: string;
}

export class InputDkTanThuLoaiKhoangSan {
  idtanthuloaikhoangsan: string;
  tenkhoangsan: string;
  idloaikhoangsan: string;
  iddangkytanthu: string;
  truluong: number;
  donvitruluong: string;
  siteid: string;
}